Refactor warn function
