+++ packages/shopify-codemod/transforms/utils.js
@@ -1,73 +1,4 @@
 import j from 'jscodeshift';
-import chalk from 'chalk';
-
-const NEWLINE_REGEX = /\r\n|[\n\v\f\r\x85\u2028\u2029]/;
-const CONTEXT_LINES = 2;
-
-function padLeft(text, length) {
-  return (' '.repeat(length) + text).slice(-length);
-}
-
-function codeContext(node, source, color) {
-  const lines = source.split(NEWLINE_REGEX);
-
-  const {start, end} = node.node.loc;
-  const startLine = start.line - 1;
-  const startColumn = start.column;
-  const endLine = end.line - 1;
-  const endColumn = end.column;
-  const contextStart = Math.max(0, startLine - CONTEXT_LINES);
-  const contextEnd = Math.min(lines.length, endLine + CONTEXT_LINES);
-
-  let result = '';
-  for (let i = contextStart; i <= contextEnd; i++) {
-    const line = lines[i];
-    let paintStart = Number.MAX_SAFE_INTEGER;
-    let paintEnd = Number.MAX_SAFE_INTEGER;
-
-    if (startLine <= i && i <= endLine) {
-      if (i === startLine) {
-        paintStart = startColumn;
-      } else {
-        paintStart = 0;
-      }
-      if (i === endLine) {
-        paintEnd = endColumn;
-      } else {
-        paintEnd = Number.MAX_SAFE_INTEGER;
-      }
-    }
-
-    result += `${chalk.gray(padLeft(i + 1, 4))}|`;
-    result += line.slice(0, paintStart);
-    result += color(line.slice(paintStart, paintEnd));
-    result += line.slice(paintEnd, Number.MAX_SAFE_INTEGER);
-    result += '\n';
-  }
-  return result;
-}
-
-function logMessage(label, message, color, file, node, source) {
-  const coloredLabel = color(`[${label}]`);
-  const fileTag = file ? ` (${file.replace(/\.coffee$/, '.js')})` : '';
-  const context = node && source ? `\n${codeContext(node, source, color)}` : '';
-  return `${coloredLabel} ${message}${fileTag}${context}`;
-}
-
-export function info(message, file, node, source) {
-  // eslint-disable-next-line no-console
-  console.info(logMessage('info', message, chalk.blue, file, node, source));
-}
-
-export function warn(message, file, node, source) {
-  // eslint-disable-next-line no-console
-  console.warn(logMessage('warning', message, chalk.yellow, file, node, source));
-}
-
-export function error(message, file, node, source) {
-  // eslint-disable-next-line no-console
-  console.error(logMessage('error', message, chalk.red, file, node, source));
-}
 
 export function findFirstMember(node) {
   if (j.MemberExpression.check(node)) {
